INTYGFV-13556: Fixed bug where empty values were dispatched from UeSickLeavePeriod
