+++ packages/common/src/utils/certificateUtils.ts
@@ -1,4 +1,4 @@
-import { ValueDateRange } from './../types/certificate';
+import { ValueDateRange } from './../types/certificate'
 import { ResourceLink } from './../types/resourceLink'
 import { CertificateEvent, CertificateMetadata, CertificateRelationType, CertificateStatus, Certificate, CertificateEventType } from '..'
 import { ResourceLinkType } from '../types/resourceLink'
@@ -75,11 +75,4 @@ export const getCertificateToSave = (certificate: Certificate): Certificate => {
     }
   }
   return cleanCertificate
-}
-
-
-export const filterDateRangeValueList = (valueList: ValueDateRange[]) => {
-  return valueList.map(
-    (val) => Object.fromEntries(Object.entries(val).filter(([_, v]) => v != null && (v as string).length > 0)) as ValueDateRange
-  )
 }
